Reduce CPU usage with a decoder that works using gpiod events.
Gpiod library also supports receiving events on rising or falling edge
for a given pin. This change adds support for that and makes it default.
The old polling mode can still be enabled by adding the following
configuration to the config file:

[receiver]
polling=1

Resource usage is significantly lower - on Raspberry Pi 2 the CPU usage
goes from 40+% to around 1%-2% user space and 1%-2% kernel IRQ that handles the
events, somewhat depending on how noisy the frequency is.